requirements.txt
fastapi
uvicorn
openai
supabase
python-dotenv
tiktoken
pdfplumber
tqdm
